update multi boot & debian sudo questions

+++ 10 - arch-post-install.sh
@@ -1,7 +1,62 @@
 #!/bin/bash
 # NOTE this script is only tested in my machines
 
-sudo rm /arch-install
+if [ -f /arch-install ]; then
+  sudo rm /arch-install
+fi
+
+while true; do
+  read -p "Will boot with other linux distros and share a partitions [yN]?   " wdb
+  case $wdb in
+    [Yy]* )
+      while true; do
+        echo "
+
+NOTE: Use a UID that will less likely be used as an ID by other distros (e.g. 1106).
+This UID will also be used on the other distro installations
+
+"
+        read -p "Enter UID or [e]xit:   " uid
+        case $uid in
+          [Ee]* ) break;;
+          * )
+            while true; do
+              echo "
+
+NOTE: Use a GUID that will less likely be used as an ID by other distros (e.g. 1106).
+This GUID will also be used on the other distro installations
+
+"
+              read -p "Enter GUID or [e]xit:   " guid
+              case $guid in
+                [Ee]* ) break 2;;
+                * )
+                  while true; do
+                    echo "
+
+Logout this user account and execute the commands below as a root user on tty2 (Ctrl + Alt + F2):
+
+groupadd wheel
+usermod -u $uid $(whoami)
+groupmod -g $guid wheel
+usermod -g wheel $(whoami)
+chown -R $(whoami):wheel /home/$(whoami)
+
+"
+                    read -p "Choose action: [l]ogout | [s]kip   " wultp
+                    case $wultp in
+                      [Ss]* ) break 4;;
+                      [Ll]* ) sudo pkill -KILL -u $(whoami);;
+                      * ) echo "Invalid input";;
+                    esac
+                  done;;
+              esac
+            done;;
+        esac
+      done;;
+    * ) break;;
+  esac
+done
 
 # https://www.archlinux.org/groups/x86_64/base-devel/
 # Current libs (3/15/2019)
